requirements.txt
Flask>=2.2.0
Flask-Login>=0.6.2
Flask-WTF>=1.0.1
Flask-Migrate>=4.0.4
Flask-SQLAlchemy>=3.0.3
psycopg2-binary>=2.9.5
python-dotenv>=1.0.0
WTForms>=3.0.1
